AsyncBackend: Added common & generic DNS resolver
